requirements.txt
dspy-ai
openai
python-dotenv
flask
requests
gunicorn